feat: Typescript migration: upper bar components
