+++ frontend/src/modules/Workplace/__tests__/test_upperbar.js
@@ -13,11 +13,10 @@
     limitations under the License.
 */
 
-import React from "react";
 import { screen, fireEvent } from "@testing-library/react";
 import { renderWithProviderAndRouter, createCategory } from "../../../utils/test-utils";
 import { initialState as initialWorkspaceState } from "../redux";
-import UpperBar from '../upperbar/UpperBar'
+import { UpperBar } from "../upperbar";
 import { categoriesExample } from '../../../utils/test-utils'
 
 test("category action butttons are present", async () => {
